pyproject.toml
[project]
name = "twipsybot"
version = "0.1.0"
description = "AI-powered Misskey bot using any OpenAI-SDK compatible LLM."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"

dependencies = [
    "aiohttp==3.13.3",
    "aiosqlite==0.22.1",
    "anyio==4.12.0",
    "apscheduler==3.11.2",
    "cachetools==6.2.4",
    "feedparser==6.0.11",
    "loguru==0.7.3",
    "openai==2.14.0",
    "psutil==7.2.1",
    "pyyaml==6.0.3",
    "python-dotenv==1.2.1",
    "tenacity==9.1.2",
]

[project.scripts]
twipsybot = "twipsybot.app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
